Request: convert to single responsibility QueryHandler
